delete unuseful classes Signed-off-by: dbulahov <bulahovdenis@gmail.com>
